utils.py updated

+++ utils.py
@@ -31,7 +31,8 @@ def pad_sequences(seqs):
     padded_seqs = torch.zeros(len(seqs), max(lens)).long()
     for i, seq in enumerate(seqs):
         end = lens[i]
-        padded_seqs[i, :end] = torch.LongTensor(seq[:end])
+        # Changed from Long to Float because: embeddings
+        padded_seqs[i, :end] = torch.FloatTensor(seq[:end])
     return padded_seqs, lens
 
 def valid_params(params):
